fix: 하트 숫자에 따른 카드 컴포넌트 UI 밀림 현상 해결
